server/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
easyocr==1.7.1
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.3.1
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
mysql-connector-python==8.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pyclipper==1.3.0.post5
python-bidi==0.4.2
PyYAML==6.0.1
scikit-image==0.22.0
scipy==1.13.0
shapely==2.0.3
six==1.16.0
sympy==1.12
tifffile==2024.2.12
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.10.0
Werkzeug==3.0.2
